+++ test/jasmine/spec/SpecHelper.js
@@ -203,3 +203,4 @@ var loadData = handsontableMethodFactory('loadData');
 var destroyEditor = handsontableMethodFactory('destroyEditor');
 var render = handsontableMethodFactory('render');
 var updateSettings = handsontableMethodFactory('updateSettings');
+var destroy = handsontableMethodFactory('destroy');